+++ cookbooks/wt_actioncenter_dd_webtrendsemaildemo/recipes/default.rb
@@ -14,45 +14,38 @@ else
   log "The deploy_build value is not set or is false so we will only update the configuration"
 end
 
-install_dir = File.join(node['wt_common']['install_dir_linux'], "harness/plugins/wt_actioncenter_dd_webtrendsemaildemo")
-conf_dir = File.join(install_dir, "conf")
+install_dir  = File.join(node['wt_portfolio_harness']['plugin_dir'], "webtrendsemaildemo")
+conf_dir     = File.join(install_dir, "conf")
 tarball      = node['wt_actioncenter_dd_webtrendsemaildemo']['download_url'].split("/")[-1]
 download_url = node['wt_actioncenter_dd_webtrendsemaildemo']['download_url']
-user = node['wt_actioncenter_dd_webtrendsemaildemo']['user']
-group = node['wt_actioncenter_dd_webtrendsemaildemo']['group']
-config_host = URI(node['wt_streamingconfigservice']['config_service_url']).host
+user         = node['wt_actioncenter_dd_webtrendsemaildemo']['user']
+group        = node['wt_actioncenter_dd_webtrendsemaildemo']['group']
+ads_host     = URI(node['wt_streamingconfigservice']['config_service_url']).host
+
+ads_ssl_port = node['wt_streamingconfigservice']['config_service_ssl_port']
+
+datarequest_max_event_batch_time_ms = node['wt_actioncenter_dd_webtrendsemaildemo']['datarequest_max_event_batch_time_ms']
+datarequest_max_events_in_batch = node['wt_actioncenter_dd_webtrendsemaildemo']['datarequest_max_events_in_batch']
+datarequest_failure_delay_before_retry_ms = node['wt_actioncenter_dd_webtrendsemaildemo']['datarequest_failure_delay_before_retry_ms']
+datarequest_nodata_delay_before_retry_ms = node['wt_actioncenter_dd_webtrendsemaildemo']['datarequest_nodata_delay_before_retry_ms']
+sender_max_send_retries = node['wt_actioncenter_dd_webtrendsemaildemo']['sender_max_send_retries']
+sender_min_exponential_backoff_delay_ms = node['wt_actioncenter_dd_webtrendsemaildemo']['sender_min_exponential_backoff_delay_ms']
+sender_max_delay_before_dropping_data_ms = node['wt_actioncenter_dd_webtrendsemaildemo']['sender_max_delay_before_dropping_data_ms']
+
 
 log "Install dir: #{install_dir}"
 
-# create the install directory
-directory "#{install_dir}" do
-  owner "root"
-  group "root"
-  mode 00755
-  recursive true
-  action :create
+# create the directories
+[install_dir, conf_dir].each do |dir|
+  directory dir do
+    owner "root"
+    group "root"
+    mode 00755
+    recursive true
+    action :create
+  end
 end
 
-directory "#{conf_dir}" do
-  owner "root"
-  group "root"
-  action :create
-end
-
-def processTemplates(conf_dir, config_host)
-	log "Updating template files"
-	%w[config.properties].each do | template_file |
-		template "#{conf_dir}/#{template_file}" do
-			source "#{template_file}.erb"
-			owner "root"
-			group "root"
-			mode 00644
-			variables({
-				:config_host => config_host,
-			})
-		end
-	end
-end
 
 if ENV["deploy_build"] == "true" then
   log "The deploy_build value is true so we will grab the tar ball and install"
@@ -64,16 +57,13 @@ if ENV["deploy_build"] == "true" then
   end
 
     # uncompress the application tarball into the install dir
-    execute "tar" do
-        user  "root"
-        group "root"
-        cwd install_dir
-        command "tar zxf #{Chef::Config[:file_cache_path]}/#{tarball}"
-    end
-
-
-	processTemplates(conf_dir, config_host)	
-
+  execute "tar" do
+    user  "root"
+    group "root"
+    cwd install_dir
+    command "tar zxf #{Chef::Config[:file_cache_path]}/#{tarball}"
+    notifies :restart, "service[harness]", :delayed
+  end
 
   # delete the install tar ball
   execute "delete_install_source" do
@@ -82,7 +72,24 @@ if ENV["deploy_build"] == "true" then
     command "rm -f #{Chef::Config[:file_cache_path]}/#{tarball}"
     action :run
   end
-
-else
 end
 
+template "#{conf_dir}/config.properties" do
+  source "config.properties.erb"
+  owner "root"
+  group "root"
+  mode 00644
+  variables({
+    :config_host => ads_host,
+    :secure_config_host => ads_host,
+    :secure_config_port => ads_ssl_port,
+    :datarequest_max_event_batch_time_ms => datarequest_max_event_batch_time_ms,
+    :datarequest_max_events_in_batch => datarequest_max_events_in_batch,
+    :datarequest_failure_delay_before_retry_ms => datarequest_failure_delay_before_retry_ms,
+    :datarequest_nodata_delay_before_retry_ms => datarequest_nodata_delay_before_retry_ms,
+    :sender_max_send_retries => sender_max_send_retries,
+    :sender_min_exponential_backoff_delay_ms => sender_min_exponential_backoff_delay_ms,
+    :sender_max_delay_before_dropping_data_ms => sender_max_delay_before_dropping_data_ms
+  })
+  notifies :restart, "service[harness]", :delayed
+end