+++ cookbooks/wt_actioncenter_management_api/metadata.rb
@@ -2,9 +2,7 @@ name                "wt_actioncenter_management_api"
 maintainer          "Webtrends, Inc."
 maintainer_email    "michael.parsons@webtrends.com"
 license             "All rights reserved"
-description         "Installs/Configures Webtrends Action Center Mangaement
-Service"
-long_description    IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version             "1.0.0"
+description         "Installs/Configures Webtrends Action Center Mangaement Service"
+version             "2.0.0"
 depends             "java"
 depends             "runit"