+++ cookbooks/ubuntu/recipes/default.rb
@@ -26,8 +26,21 @@ template "/etc/apt/sources.list" do
     :code_name => node['lsb']['codename'],
     :security_url => node['ubuntu']['security_url'],
     :archive_url => node['ubuntu']['archive_url'],
-    :include_source_packages => node['ubuntu']['include_source_packages']
+    :include_source_packages => node['ubuntu']['include_source_packages'],
+    :components => node['ubuntu']['components']
   )
   notifies :run, "execute[apt-get update]", :immediately
   source "sources.list.erb"
+end
+
+if node['ubuntu']['locale']
+
+ [ "LC_ALL", "LANG" ].each do |envvar|
+    execute "set_locale_#{envvar.downcase}" do
+      command "update-locale #{envvar}=#{node['ubuntu']['locale']}"
+      action :run
+      not_if "grep #{envvar}=#{node['ubuntu']['locale']} /etc/default/locale"
+    end
+  end
+
 end