+++ cookbooks/wt_actioncenter_ddp/metadata.rb
@@ -2,9 +2,7 @@ name                "wt_actioncenter_ddp"
 maintainer          "Webtrends, Inc."
 maintainer_email    "michael.parsons@webtrends.com"
 license             "All rights reserved"
-description         "Installs/Configures Webtrends Action Center Data
-Destination Processor Service"
-long_description    IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version             "1.0.0"
+description         "Installs/Configures Webtrends Action Center Data Destination Processor Service"
+version             "1.1.3"
 depends             "java"
 depends             "runit"