+++ cookbooks/wt_streamingcollection/CHANGELOG.md
@@ -1,5 +1,6 @@
+
 ## 1.1.7
-* Change config service healthcheck to /ping instead of /whitelist/collection
+* Updated the Nagios check to make sure it calls the health check on the main port (i.e. 8080)
 
 ## 1.1.6
 * Cleanup runit service disabling in undeploy.rb